Adiciona lógica para permitir acesso as atividades apenas aos usários que às criaram

+++ src/components/activity/activity.controller.ts
@@ -14,7 +14,8 @@ export class ActivityController extends ActivityService implements IActivityCont
 
     public async createActivity(req: Request, res: Response): Promise<void> {
         try {
-            const activity: IStoreActivityDTO = req.body;
+            const userID: number = Number(req.headers['user-id'])
+            const activity: IStoreActivityDTO = {...req.body, userID};
 
             if(activity) {
                 const activityExists: IResponseJson = await super.findExistingNameActivityService(activity.name);
@@ -46,29 +47,38 @@ export class ActivityController extends ActivityService implements IActivityCont
     public async deleteActivity(req: Request, res: Response): Promise<void> {
         try {
             const id: number = Number(req.params.id);
+            const userID: number = Number(req.headers['user-id']);
+            const permission: IResponseJson = await super.compareActivityIdAndUserID(id, userID);
             
-            if(id) {
-                const idExists: IResponseJson = await super.findByIdActivityService(id);
-    
-                if(idExists.status && idExists.data) {
-                    const result: IResponseJson = await super.deleteActivityService(id);
-            
-                    if(result.status) {
-                        res.status(200);
-                        res.send("Activity deleted successfully!");
+            if(permission.data) {
+                if(id) {
+                    const idExists: IResponseJson = await super.findByIdActivityService(id);
+        
+                    if(idExists.status && idExists.data) {
+                        const result: IResponseJson = await super.deleteActivityService(id);
+                
+                        if(result.status) {
+                            res.status(200);
+                            res.send("Activity deleted successfully!");
+                        } else {
+                            res.status(400);
+                            res.json(result?.message);
+                        }
+                    } else if(idExists.status && !idExists.data) {
+                        res.status(404);
+                        res.send("Activity not found!")
+                        throw Error("Activity not found!") 
                     } else {
-                        res.status(400);
-                        res.json(result?.message);
+                        res.status(400)
+                        res.send(idExists?.message);
                     }
-                } else if(idExists.status && !idExists.data) {
-                    res.status(404);
-                    res.send("Activity not found!")
-                    throw Error("Activity not found!") 
-                } else {
-                    res.status(400)
-                    res.send(idExists?.message);
                 }
+            } else {
+                res.status(401);
+                res.json("Not authorized!");
+                throw Error("Not authorized!");
             }
+
         } catch(error: any) {
             console.error(error);
         }
@@ -77,21 +87,29 @@ export class ActivityController extends ActivityService implements IActivityCont
     public async findActivity(req: Request, res: Response): Promise<void> {
         try {
             const id: number = Number(req.params.id);
+            const userID: number = Number(req.headers['user-id']);
+            const permission: IResponseJson = await super.compareActivityIdAndUserID(id, userID);
 
-            if(id) {
-                const result: IResponseJson = await super.findByIdActivityService(id);
-
-                if(result.status && result.data) {
-                    res.status(200);
-                    res.json({status: 200, data: result.data});
-                } else if(result.status && !result.data) {
-                    res.status(404);
-                    res.json("Activity not found!");
-                    throw Error("Activity not found!") 
-                } else {
-                    res.status(400);
-                    res.json(result?.message);
+            if(permission.data) {
+                if(id) {
+                    const result: IResponseJson = await super.findByIdActivityService(id);
+    
+                    if(result.status && result.data) {
+                        res.status(200);
+                        res.json({status: 200, data: result.data});
+                    } else if(result.status && !result.data) {
+                        res.status(404);
+                        res.json("Activity not found!");
+                        throw Error("Activity not found!") 
+                    } else {
+                        res.status(400);
+                        res.json(result?.message);
+                    }
                 }
+            } else {
+                res.status(401);
+                res.json("Not authorized!");
+                throw Error("Not authorized!");
             }
         } catch(error: any) {
             console.error(error);
@@ -100,15 +118,12 @@ export class ActivityController extends ActivityService implements IActivityCont
 
     public async findAllActivities(req: Request, res: Response): Promise<void> {
         try {
-            const result: IResponseJson = await super.findAllActivityService();
+            const userID: number = Number(req.headers['user-id'])
+            const result: IResponseJson = await super.findAllActivityService(userID);
 
-            if(result.status && result.data) {
+            if(result.status) {
                 res.status(200);
                 res.json({status: 200, data: result.data});
-            } else if(result.status && !result.data) {
-                res.status(404);
-                res.json("Activities not found!");
-                throw Error("Activities not found!") 
             } else {
                 res.status(400);
                 res.json(result?.message);
@@ -120,30 +135,39 @@ export class ActivityController extends ActivityService implements IActivityCont
 
     public async updateActivity(req: Request, res: Response): Promise<void> {
         try {
-            const activity: IUpdateActivityDTO = req.body;
+            const id: number = Number(req.params.id);
+            const activity: IUpdateActivityDTO = {...req.body, id};
+            const userID: number = Number(req.headers['user-id']);
+            const permission: IResponseJson = await super.compareActivityIdAndUserID(activity.id, userID);
 
-            if(activity) {
-                const idExists: IResponseJson = await super.findByIdActivityService(activity.id);
-    
-                if(idExists.status && idExists.data) {
-                   
-                    const result: IResponseJson = await super.updateActivityService(activity);
+            if(permission.data) {
+                if(activity) {
+                    const idExists: IResponseJson = await super.findByIdActivityService(activity.id);
         
-                    if(result.status) {
-                        res.status(200);
-                        res.json({success:1});
+                    if(idExists.status && idExists.data) {
+                       
+                        const result: IResponseJson = await super.updateActivityService(activity);
+            
+                        if(result.status) {
+                            res.status(200);
+                            res.json({success:1});
+                        } else {
+                            res.status(400);
+                            res.json(result?.message);
+                        }
+                    } else if(idExists.status && !idExists.data) {
+                        res.status(404);
+                        res.json("Activity not found!");
+                        throw Error("Activity not found!") 
                     } else {
                         res.status(400);
-                        res.json(result?.message);
+                        res.json(idExists?.message);
                     }
-                } else if(idExists.status && !idExists.data) {
-                    res.status(404);
-                    res.json("Activity not found!");
-                    throw Error("Activity not found!") 
-                } else {
-                    res.status(400);
-                    res.json(idExists?.message);
                 }
+            } else {
+                res.status(401);
+                res.json("Not authorized!");
+                throw Error("Not authorized!");
             }
         } catch (error: any) {
             console.error(error);
